Support static values in the RT to CSV field mapping
Static values are indicated by mapping the RT column to a string ref.

This is primarily useful for setting a Catalog for every record in the
import without modifying the CSV.  However, it may also be useful when
the CSV generation is not controlled by the user importing and static
values can eliminate a preprocessing step.