Allow importing with a numeric unique column used for asset ids
This allows RT to take ownership of an pre-existing serial number during
import.  While it requires that such an import only happen once (upon
which RT becomes blessed as the generator of all future asset IDs), it
allows RT's pervasive concept of "asset number" to line up with users'
pre-existing concepts of such when moving from an existing system.

After importing, the sequences must be updated soas to not overlap with
existing ids.  This is unnecessary on mysql and SQLite, as they adjust
their auto_increment values upon seeing the insert, but requires
explicit steps on PostgreSQL and Oracle.