Allow import into multi-value role groups
Previously you could only set your Asset Owners, not Held By or Contact
during the import.  You still can't set two Held Bys during the import,
because we haven't defined a convention for what would go in the field
(a comma separated CSV field).

Perhaps switching to allowing email addresses and parsing for them, but
we currently use Username (or Id), and don't deal with email addresses
at all so that would be a change.