Add some sanity checks to the CSV library
Add sanity checks to the CSV "run" method and return wanings when the
issues are found.

Check if the columns in the CSV file are being used by the mapping or
if there are declared fields in the mapping that are not present in the
CSV file.

Some fields can be declared inside a CODE reference mapping. That's why
we added Data::Dumper to evaluate those CODE refence fields and check
which fields are present inside the code block. That way we can also add
those fields to the comparison with the CSV headers.